Moved utility modules to subdirectory to improve organization
