+++ import/util/importutility.py
@@ -1,16 +1,16 @@
 #!/usr/bin/env python
 """Utility methods for import.py"""
-from dbconnect import connect
-import find
+import util.find
 from datetime import datetime as dt
 import pandas as pd
 import numpy as np
 import psycopg2
 import csv
-import insert
-from dbconnect import config, connect
-from models import species, population, line, chromosome, variant, genotype, trait, phenotype, growout_type, growout, location, gwas_algorithm, genotype_version, imputation_method, kinship_algorithm, kinship, population_structure_algorithm, population_structure, gwas_run, gwas_result
-
+import util.insert
+from util.dbconnect import config, connect
+from util.models import species, population, line, chromosome, variant, genotype, trait, phenotype, growout_type, growout, location, gwas_algorithm, genotype_version, imputation_method, kinship_algorithm, kinship, population_structure_algorithm, population_structure, gwas_run, gwas_result
+import util.validate as validate
+import util.find as find
 
 def growout_name_to_year(value):
   """Takes in a LOYR (Location-Year) string and converts year digits to four-digit year"""
