--- modules/database.ini
@@ -1,6 +1,0 @@
-[postgresql]
-database=baxdb
-user=baxdb_owner
-password=password
-host=localhost
-port=5434
